Allow schemas to be used in VS Code
The reference mechanism used for the schemas did not actually match any
real online resources, which throws off the JSON language support in VS
Code. As each object isn't actually used more than once in each schema,
we might as well inline them all and simplify both the resulting schemas
and the generation script.

Additionally sets the schema for index files in the vscode workspace
settings, so that they're automatically validated when editing.