Add license to app listing (#26)
* Add license to app listing

Adds a small license icon and name at the bottom left of each AppBlock.
This primarily relies on GitHub's license detection mechanism, but can
be overwritten in the application entries if GitHub fails to detect the
right one.

* Add manual license entry in Golioth's apps

GitHub's license detection fails in two of Golioth's apps, but based on
the content, it's meant to be Apache 2.0.