Add GitHub stats (#27)
* Replace heroicons with Octicons

Heroicons don't match the octicon look for stars, forks and watchers,
but Octicons have equivalents for all the icons we're using from
heroicons, so we can replace them all.

* Pull stats from GitHub

Shows stars, watchers and forks next to the license text on each app
block. Changes the app sorting to star count.

* Schedule publication to run every night

Updates the stats for each repo.